requirements.txt
SQLAlchemy==2.0.10
mysql-connector-python==8.0.33
python-dotenv==1.0.0